New folder/backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
httpx 
pytesseract
pillow 
pdf2image 